move pause bit verify to solidity
