+++ builtin/native_calls_test.go
@@ -12,20 +12,16 @@ import (
 	"fmt"
 	"math"
 	"math/big"
-	"reflect"
 	"testing"
-	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/vechain/thor/v2/abi"
-	"github.com/vechain/thor/v2/block"
 	"github.com/vechain/thor/v2/builtin"
 	"github.com/vechain/thor/v2/builtin/staker/validation"
 	"github.com/vechain/thor/v2/genesis"
-	"github.com/vechain/thor/v2/muxdb"
 	"github.com/vechain/thor/v2/runtime"
 	"github.com/vechain/thor/v2/state"
 	"github.com/vechain/thor/v2/test/testchain"
@@ -33,40 +29,10 @@ import (
 	_ "github.com/vechain/thor/v2/tracers/native"
 	"github.com/vechain/thor/v2/trie"
 	"github.com/vechain/thor/v2/tx"
-	"github.com/vechain/thor/v2/vm"
 	"github.com/vechain/thor/v2/xenv"
 )
 
-var (
-	errReverted = vm.ErrExecutionReverted
-	revertABI   = []byte(`[{"name": "Error","type": "function","inputs": [{"name": "message","type": "string"}]}]`)
-	thorChain   *testchain.Chain
-)
-
-type ctest struct {
-	rt         *runtime.Runtime
-	abi        *abi.ABI
-	to, caller thor.Address
-}
-
-type ccase struct {
-	rt         *runtime.Runtime
-	abi        *abi.ABI
-	to, caller thor.Address
-	name       string
-	args       []any
-	events     tx.Events
-	provedWork *big.Int
-	txID       thor.Bytes32
-	blockRef   tx.BlockRef
-	gasPayer   thor.Address
-	expiration uint32
-	value      *big.Int
-
-	output    *[]any
-	vmerr     error
-	revertMsg string
-}
+var thorChain *testchain.Chain
 
 type TestTxDescription struct {
 	t          *testing.T
@@ -77,167 +43,6 @@ type TestTxDescription struct {
 	args       []any
 	duplicate  bool
 	vet        *big.Int
-}
-
-func (c *ctest) Case(name string, args ...any) *ccase {
-	return &ccase{
-		rt:     c.rt,
-		abi:    c.abi,
-		to:     c.to,
-		caller: c.caller,
-		name:   name,
-		args:   args,
-	}
-}
-
-func (c *ccase) To(to thor.Address) *ccase {
-	c.to = to
-	return c
-}
-
-func (c *ccase) Caller(caller thor.Address) *ccase {
-	c.caller = caller
-	return c
-}
-
-func (c *ccase) Value(value *big.Int) *ccase {
-	c.value = value
-	return c
-}
-
-func (c *ccase) ProvedWork(provedWork *big.Int) *ccase {
-	c.provedWork = provedWork
-	return c
-}
-
-func (c *ccase) TxID(txID thor.Bytes32) *ccase {
-	c.txID = txID
-	return c
-}
-
-func (c *ccase) BlockRef(blockRef tx.BlockRef) *ccase {
-	c.blockRef = blockRef
-	return c
-}
-
-func (c *ccase) GasPayer(gasPayer thor.Address) *ccase {
-	c.gasPayer = gasPayer
-	return c
-}
-
-func (c *ccase) Expiration(expiration uint32) *ccase {
-	c.expiration = expiration
-	return c
-}
-
-func (c *ccase) ShouldVMError(err error) *ccase {
-	c.vmerr = err
-	return c
-}
-
-func (c *ccase) ShouldLog(events ...*tx.Event) *ccase {
-	c.events = events
-	return c
-}
-
-func (c *ccase) ShouldOutput(outputs ...any) *ccase {
-	c.output = &outputs
-	return c
-}
-
-func (c *ccase) ShouldRevert(revertMsg string) *ccase {
-	c.revertMsg = revertMsg
-	c.vmerr = errReverted
-	return c
-}
-
-func (c *ccase) Assert(t *testing.T) *ccase {
-	method, ok := c.abi.MethodByName(c.name)
-	assert.True(t, ok, "should have method")
-
-	constant := method.Const()
-	stage, err := c.rt.State().Stage(trie.Version{})
-	assert.Nil(t, err, "should stage state")
-	stateRoot := stage.Hash()
-
-	data, err := method.EncodeInput(c.args...)
-	assert.Nil(t, err, "should encode input")
-
-	clause := tx.NewClause(&c.to).WithData(data)
-	if c.value != nil {
-		clause = clause.WithValue(c.value)
-	}
-
-	exec, _ := c.rt.PrepareClause(clause,
-		0, 40000000, &xenv.TransactionContext{
-			ID:         c.txID,
-			Origin:     c.caller,
-			GasPrice:   &big.Int{},
-			GasPayer:   c.gasPayer,
-			ProvedWork: c.provedWork,
-			BlockRef:   c.blockRef,
-			Expiration: c.expiration,
-		})
-	vmout, _, err := exec()
-	assert.Nil(t, err)
-	if constant || vmout.VMErr != nil {
-		stage, err := c.rt.State().Stage(trie.Version{})
-		assert.Nil(t, err, "should stage state")
-		newStateRoot := stage.Hash()
-		assert.Equal(t, stateRoot, newStateRoot)
-	}
-	if c.vmerr != nil {
-		assert.Equal(t, c.vmerr, vmout.VMErr)
-	} else {
-		assert.Nil(t, vmout.VMErr)
-	}
-
-	if c.output != nil {
-		out, err := method.EncodeOutput((*c.output)...)
-		assert.Nil(t, err, "should encode output")
-		assert.Equal(t, out, vmout.Data, "should match output")
-	}
-
-	if len(c.events) > 0 {
-		for _, ev := range c.events {
-			found := func() bool {
-				for _, outEv := range vmout.Events {
-					if reflect.DeepEqual(ev, outEv) {
-						return true
-					}
-				}
-				return false
-			}()
-			assert.True(t, found, "event should appear")
-		}
-	}
-
-	if c.revertMsg != "" {
-		abis, err := abi.New(revertABI)
-		assert.NoError(t, err)
-		method, ok := abis.MethodByName("Error")
-		assert.True(t, ok)
-		var revertMsg string
-		err = method.DecodeInput(vmout.Data, &revertMsg)
-		assert.NoError(t, err)
-		assert.Equal(t, c.revertMsg, revertMsg)
-	}
-
-	c.output = nil
-	c.vmerr = nil
-	c.events = nil
-	c.revertMsg = ""
-
-	return c
-}
-
-func buildGenesis(db *muxdb.MuxDB, proc func(state *state.State) error) *block.Block {
-	blk, _, _, _ := new(genesis.Builder).
-		Timestamp(uint64(time.Now().Unix())).
-		State(proc).
-		ForkConfig(&thor.NoFork).
-		Build(state.NewStater(db))
-	return blk
 }
 
 func inspectClauseWithBlockRef(clause *tx.Clause, blockRef *tx.BlockRef) ([]byte, uint64, error) {
@@ -777,7 +582,7 @@ func TestEnergyNative(t *testing.T) {
 
 	assert.NoError(t, params.MintTransaction("set", big.NewInt(0), thor.KeyMaxBlockProposers, big.NewInt(1)))
 	shouldCollectEnergy()
-	assert.NoError(t, params.MintTransaction("set", big.NewInt(0), thor.KeyStargateContractAddress, big.NewInt(0).SetBytes(acc4.Bytes())))
+	assert.NoError(t, params.MintTransaction("set", big.NewInt(0), thor.KeyDelegatorContractAddress, big.NewInt(0).SetBytes(acc4.Bytes())))
 	shouldCollectEnergy()
 
 	// 2: Add a validator to the queue
@@ -2053,86 +1858,4 @@ func TestExtensionV3(t *testing.T) {
 	assert.Nil(t, err)
 	val = new(big.Int).SetBytes(out.Data)
 	assert.Equal(t, uint64(clauseCount), val.Uint64())
-}
-
-func TestStakerContract_Native_CheckStake(t *testing.T) {
-	var (
-		caller     = genesis.DevAccounts()[0].Address
-		master     = thor.BytesToAddress([]byte("master"))
-		validation = thor.BytesToBytes32([]byte("validation"))
-		delegator  = thor.Address{}
-	)
-
-	fc := &thor.SoloFork
-	fc.HAYABUSA = 1
-	fc.HAYABUSA_TP = 2
-	var err error
-	thorChain, err = testchain.NewWithFork(fc, 180)
-	assert.NoError(t, err)
-	assert.NoError(t, thorChain.MintBlock(genesis.DevAccounts()[0]))
-	assert.NoError(t, thorChain.MintBlock(genesis.DevAccounts()[0]))
-
-	thorChain.MintClauses(genesis.DevAccounts()[0], []*tx.Clause{
-		tx.NewClause(&delegator).WithValue(big.NewInt(1e18)),
-	})
-
-	rt := runtime.New(
-		thorChain.Repo().NewBestChain(),
-		thorChain.Stater().NewState(thorChain.Repo().BestBlockSummary().Root()),
-		&xenv.BlockContext{Time: thorChain.Repo().BestBlockSummary().Header.Timestamp()},
-		thorChain.GetForkConfig(),
-	)
-
-	test := &ctest{
-		rt:     rt,
-		abi:    builtin.Staker.ABI,
-		to:     builtin.Staker.Address,
-		caller: builtin.Staker.Address,
-	}
-
-	test.Case("addValidation", master, thor.LowStakingPeriod()).
-		Value(big.NewInt(0)).
-		Caller(caller).
-		ShouldRevert("stake is empty").
-		Assert(t)
-
-	test.Case("addValidation", master, thor.LowStakingPeriod()).
-		Value(big.NewInt(1)).
-		Caller(caller).
-		ShouldRevert("stake is not multiple of 1VET").
-		Assert(t)
-
-	test.Case("increaseStake", validation).
-		Value(big.NewInt(0)).
-		Caller(caller).
-		ShouldRevert("stake is empty").
-		Assert(t)
-
-	test.Case("increaseStake", validation).
-		Value(big.NewInt(1)).
-		Caller(caller).
-		ShouldRevert("stake is not multiple of 1VET").
-		Assert(t)
-
-	test.Case("decreaseStake", validation, big.NewInt(0)).
-		Caller(caller).
-		ShouldRevert("stake is empty").
-		Assert(t)
-
-	test.Case("decreaseStake", validation, big.NewInt(1)).
-		Caller(caller).
-		ShouldRevert("stake is not multiple of 1VET").
-		Assert(t)
-
-	test.Case("addDelegation", validation, uint8(100)).
-		Caller(delegator).
-		Value(big.NewInt(0)).
-		ShouldRevert("stake is empty").
-		Assert(t)
-
-	test.Case("addDelegation", validation, uint8(100)).
-		Caller(delegator).
-		Value(big.NewInt(1)).
-		ShouldRevert("stake is not multiple of 1VET").
-		Assert(t)
 }
